fixed speed change on length

+++ Slider_Control_V3/Slider_Control_V3.ino
@@ -215,7 +215,7 @@ static const unsigned char PROGMEM image_data_Logosmall[] = {
 
   
 // Stored Values
-  float LENGTH = length_Inches * Steps_Per_inch;  //converts the user set length into usable steps
+  unsigned long LENGTH = length_Inches * Steps_Per_inch;  //converts the user set length into usable steps
   bool changeValue = false;                       //determines whether to configure the operation values or scroll down a menu, this is flipped by clicking encoder while over value
   int changeCounter = 0;                          //if change value is true "changeCounter" will be changed instead of the normal "counter" value
   bool runningPath = false;                       //tracks wether or not an operation is curretly running
